Polishing and Refactoring
Polished and refactored both views. Added more relevant info for movie_details. Connectivity_plus package not implemented for lack of time.

This sums up the prototype development. It contemplates the main features necessary for the API. Next step would be bringing the prototype to the design team vision of the product.