Refactoring and Template class for LoadingAnimationWithTimeout
Right now there is no timeout check when loading movie images with cachedNetworkImageProvider.

This makes it so that if there is no internet connection active, cachedNetworkImageProvider will keep waiting for a response, and showing the  CircularProgressIndicator.

LoadingAnimationWithTimeout is a prototype function to implement a timer and stop the loading animation after a set time.

However the clean solution would be to first check if there is an internet connection active (through the connectivity_plus package), and than decide whether to call CachedNetworkImagfeProvider and fetch the image online or to try and fetch it locally, and display a template image in case of failure.

Unfortunately there is not enough time to finish either solution.